Linked List with all functions
In this code i wrote all the functions of linked list in a single program. In this program, all the functions are called by using the switch case statement. first of all you have to create a linked list and than as you entered the choice number according the number function is called and functionality is started.